// kilakitu_temp/src/app/model/property.ts
export interface Property {
  id: number;
  title: string;
  type: string;
  location: string;
  guests: number;
  bathrooms: number;
  cost: number;
  waterdeposit: number;
  beds: number;
  zone: string;
  country: string;
  county: string;
  address: string;
  paymentdetails: string;
  description: string;
  status: string;
  images: string;
  amenities: string;
  extras: string;
  img: string;
}